pyproject.toml
[build-system]
requires = ["setuptools>=80.9.0", "setuptools-scm>=9.2", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "medrap"
dynamic = ["version"]
authors = [
  {name="Matthew McDermott", email="mattmcdermott8@gmail.com"},
  {name="Egill A. Fridgeirsson", email="egillax@gmail.com"},
]
description = "Retrieval-augmented pretraining components for MEDS-style EHR data"
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  "hydra-core",
  "hydra-zen",
  "meds-torch-data",
]

[dependency-groups]
dev = [
  "pre-commit<4", "ruff", "pytest", "pytest-cov", "pretty-print-directory"
]
docs = []

[tool.setuptools_scm]

[tool.setuptools.package-data]
medrap = ["conf/**/*.yaml"]

[project.optional-dependencies]

[project.scripts]
medrap = "medrap.cli:main"

[tool.pytest.ini_options]
addopts = [
  "--color=yes",
  "--doctest-modules",
  "--ignore=docs",
  "--ignore=AGENTS.md",
  "--ignore=CONTRIBUTORS.md",
  "--doctest-glob=*.md",
]
markers = []
doctest_optionflags = ["NORMALIZE_WHITESPACE", "ELLIPSIS"]

[project.urls]
Homepage = "https://github.com/McDermottHealthAI/MedRAP"
Issues = "https://github.com/McDermottHealthAI/MedRAP/issues"

[tool.ruff]
target-version = "py312"
line-length = 110
extend-exclude = ["docs/index.md"]

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "UP",
    "B",
    "C4",
    "DTZ",
    "RUF",
    "SIM",
    "TCH",
    "N",
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"test_*" = ["DTZ001"]

[tool.ruff.lint.pep8-naming]
extend-ignore-names = ["*MEDS*", "*NRT*", "D"]
